Lots of good refactoring work!
Heatmap and boxplot can now be made in a snap. Might be a good idea to find a way to cache the values so I don't have to recompute them everytime the user needs or wants them.

Either way, we'll consider the exploratory thing done for now and I can move on to the contrast matrix...

AFTER it's identified that a new test set works just as well as what I've got so far